Renamed WCFeatureTooltip to WCPromoTooltip
